+++ app/src/main/java/com/example/gh_coursework/domain/usecase/point_preview/DeletePointUseCase.kt
@@ -1,5 +1,5 @@
 package com.example.gh_coursework.domain.usecase.point_preview
 
 interface DeletePointUseCase {
-    suspend fun invoke(pointId: Int)
+    suspend fun invoke(pointId: Long)
 }
